Delete TableFull Style and merge it with Table Style
